+++ app/lib/model/level6_9/Favourites.scala
@@ -1,4 +1,4 @@
-package lib.model.level9
+package lib.model.level6_9
 
 import ixias.model._
 import java.time.LocalDateTime
