+++ app/lib/model/level6_9/Purchasing.scala
@@ -1,4 +1,4 @@
-package lib.model.level6_7
+package lib.model.level6_9
 
 import ixias.model._
 import java.time.{ LocalDate, LocalDateTime }
