+++ app/lib/model/level6_9/Report.scala
@@ -1,4 +1,4 @@
-package lib.model.level8
+package lib.model.level6_9
 
 import ixias.model._
 import java.time.LocalDateTime
